pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claw-turbo"
version = "0.1.0"
description = "Zero-latency, zero-ML skill routing middleware for OpenClaw"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0"]

[project.scripts]
claw-turbo = "claw_turbo.cli:main"

[tool.setuptools.packages.find]
include = ["claw_turbo*"]
